{{ cookiecutter.project_slug }}/requirements/base.txt
# Base
typing_extensions==4.2.0
pytz==2022.1
click==8.1.3
python-slugify==6.1.2
python-multipart==0.0.5
python-environ==0.4.54

# FastAPI
fastapi==0.78.0
starlette==0.19.1
pydantic==1.9.1
h11==0.13.0
idna==3.3

# Async
anyio==3.6.1
sniffio==1.2.0
# aiofiles==0.8.0 # uncomment if you want to use files asynchronously in your app

# Uvicorn
uvicorn==0.18.2

{% if cookiecutter.database == "Postgres" %}
# Postgres
psycopg2-binary==2.9.3
{% endif %}

{% if cookiecutter.cloud_provider == "AWS" %}
# AWS
botocore==1.27.21
boto3==1.24.21
{% endif %}
